Исправлены замечания по типу данных переменной kol(количество), тип данных изменен на double

+++ src/converter.java
@@ -15,7 +15,7 @@ public class converter {
             if (mas>=1 && mas<=4 ) {
             Scanner in2 = new Scanner(System.in);
             System.out.println("Введите число: ");
-            int kol=in2.nextInt(); // в kol кладем число - количество, выбранной единицы измерения
+            double kol=in2.nextDouble(); // в kol кладем число - количество, выбранной единицы измерения
             if(kol>=0) { //Проверка на валидность kol
                 System.out.println("Результат:");
                 switch (mas) {
@@ -66,7 +66,7 @@ public class converter {
             if (ras >= 1 && ras <= 4) {
                 Scanner in2 = new Scanner(System.in);
                 System.out.println("Введите число: ");
-                int kol = in2.nextInt(); // в kol кладем число - количество, выбранной единицы измерения
+                double kol = in2.nextDouble(); // в kol кладем число - количество, выбранной единицы измерения
                 if (kol >= 0) { //Проверка на валидность kol
                     System.out.println("Результат:");
                     switch (ras) {
